refactor(core): extract punctuation window tiling
Move WINDOW_* constants, word_spans, plan_windows, and
splice_window_labels into punctuation/windows.rs. Public
Punctuator / process_token paths are unchanged.
