+++ crates/gigastt-core/src/punctuation/tests.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use super::windows::{WINDOW_OVERLAP_WORDS, WINDOW_WORDS, Window};
 use super::*;
 use crate::runtime::{RuntimeError, factory::Runtime};
